backend/requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
SQLAlchemy==1.4.46
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
python-dotenv==0.19.1
pytest==6.2.5
qrcode==7.3.1
Pillow>=9.2.0
gunicorn==20.1.0
requests==2.28.1 